@@ -1,30 +1,30 @@
 package org.snowcorp.sample.uploadfiles;
 
 import android.Manifest;
-import android.content.ActivityNotFoundException;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
-import android.widget.Button;
 import android.widget.ListView;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+
+import com.google.android.material.button.MaterialButton;
+import com.google.android.material.snackbar.Snackbar;
+
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 import okhttp3.MediaType;
 import okhttp3.MultipartBody;
@@ -38,78 +38,64 @@ import retrofit2.converter.gson.GsonConverterFactory;
 
 public class MainActivity extends AppCompatActivity {
     private static final String TAG = MainActivity.class.getSimpleName();
-    private static final int REQUEST_CODE = 6384;
-    private static final int REQUEST_CODE_ASK_PERMISSIONS = 124;
 
-    private View parentView;
     private ListView listView;
     private ProgressBar mProgressBar;
-    private Button btnChoose, btnUpload;
+    private MaterialButton btnChoose, btnUpload;
 
     private ArrayList<Uri> arrayList;
+
+    private final int REQUEST_CODE_PERMISSIONS  = 1;
+    private final int REQUEST_CODE_READ_STORAGE = 2;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        parentView = findViewById(R.id.parent_layout);
         listView = findViewById(R.id.listView);
         mProgressBar = findViewById(R.id.progressBar);
 
         btnChoose = findViewById(R.id.btnChoose);
-        btnChoose.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                // Display the file chooser dialog
-                if (askForPermission())
-                    showChooser();
+        btnChoose.setOnClickListener(v -> {
+            // Display the file chooser dialog
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                askForPermission();
+            } else {
+                showChooser();
             }
         });
 
         btnUpload = findViewById(R.id.btnUpload);
-        btnUpload.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                uploadImagesToServer();
-            }
-        });
+        btnUpload.setOnClickListener(v -> uploadImagesToServer());
 
         arrayList = new ArrayList<>();
     }
 
     private void showChooser() {
-        // Use the GET_CONTENT intent from the utility class
-        Intent target = FileUtils.createGetContentIntent();
-        // Create the chooser Intent
-        Intent intent = Intent.createChooser(
-                target, getString(R.string.chooser_title));
-        try {
-            startActivityForResult(intent, REQUEST_CODE);
-        } catch (ActivityNotFoundException e) {
-            // The reason for the existence of aFileChooser
-        }
+        Intent intent = new Intent(Intent.ACTION_GET_CONTENT);
+        intent.setType("image/*");
+        intent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true);
+        intent.addCategory(Intent.CATEGORY_OPENABLE);
+        startActivityForResult(intent, REQUEST_CODE_READ_STORAGE);
     }
 
     @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        switch (requestCode) {
-            case REQUEST_CODE:
-                // If the file selection was successful
-                if (resultCode == RESULT_OK) {
-                    if(data.getClipData() != null) {
-                        int count = data.getClipData().getItemCount();
+    protected void onActivityResult(int requestCode, int resultCode, Intent resultData) {
+        super.onActivityResult(requestCode, resultCode, resultData);
+        if (resultCode == RESULT_OK) {
+            if (requestCode == REQUEST_CODE_READ_STORAGE) {
+                if (resultData != null) {
+                    if (resultData.getClipData() != null) {
+                        int count = resultData.getClipData().getItemCount();
                         int currentItem = 0;
-                        while(currentItem < count) {
-                            Uri imageUri = data.getClipData().getItemAt(currentItem).getUri();
-                            //do something with the image (save it to some directory or whatever you need to do with it here)
+                        while (currentItem < count) {
+                            Uri imageUri = resultData.getClipData().getItemAt(currentItem).getUri();
                             currentItem = currentItem + 1;
-                            Log.d("Uri Selected", imageUri.toString());
-                            try {
-                                // Get the file path from the URI
-                                String path = FileUtils.getPath(this, imageUri);
-                                Log.d("Multiple File Selected", path);
 
+                            Log.d("Uri Selected", imageUri.toString());
+
+                            try {
                                 arrayList.add(imageUri);
                                 MyAdapter mAdapter = new MyAdapter(MainActivity.this, arrayList);
                                 listView.setAdapter(mAdapter);
@@ -118,15 +104,12 @@ public class MainActivity extends AppCompatActivity {
                                 Log.e(TAG, "File select error", e);
                             }
                         }
-                    } else if(data.getData() != null) {
-                        //do something with the image (save it to some directory or whatever you need to do with it here)
-                        final Uri uri = data.getData();
-                        Log.i(TAG, "Uri = " + uri.toString());
-                        try {
-                            // Get the file path from the URI
-                            final String path = FileUtils.getPath(this, uri);
-                            Log.d("Single File Selected", path);
+                    } else if (resultData.getData() != null) {
 
+                        final Uri uri = resultData.getData();
+                        Log.i(TAG, "Uri = " + uri.toString());
+
+                        try {
                             arrayList.add(uri);
                             MyAdapter mAdapter = new MyAdapter(MainActivity.this, arrayList);
                             listView.setAdapter(mAdapter);
@@ -136,15 +119,14 @@ public class MainActivity extends AppCompatActivity {
                         }
                     }
                 }
-                break;
+            }
         }
-        super.onActivityResult(requestCode, resultCode, data);
     }
 
     private void uploadImagesToServer() {
         if (InternetConnection.checkConnection(MainActivity.this)) {
             Retrofit retrofit = new Retrofit.Builder()
-                    .baseUrl("http://192.168.43.166/~snow/upload-files/")
+                    .baseUrl(ApiService.BASE_URL)
                     .addConverterFactory(GsonConverterFactory.create())
                     .build();
 
@@ -178,14 +160,17 @@ public class MainActivity extends AppCompatActivity {
                         Toast.makeText(MainActivity.this,
                                 "Images successfully uploaded!", Toast.LENGTH_SHORT).show();
                     } else {
-                        Snackbar.make(parentView, R.string.string_some_thing_wrong, Snackbar.LENGTH_LONG).show();
+                        Snackbar.make(findViewById(android.R.id.content),
+                                R.string.string_some_thing_wrong, Snackbar.LENGTH_LONG).show();
                     }
                 }
 
                 @Override
                 public void onFailure(@NonNull Call<ResponseBody> call, @NonNull Throwable t) {
                     hideProgress();
-                    Snackbar.make(parentView, t.getMessage(), Snackbar.LENGTH_LONG).show();
+                    Log.e(TAG, "Image upload failed!", t);
+                    Snackbar.make(findViewById(android.R.id.content),
+                            "Image upload failed!", Snackbar.LENGTH_LONG).show();
                 }
             });
 
@@ -199,33 +184,27 @@ public class MainActivity extends AppCompatActivity {
     private void showProgress() {
         mProgressBar.setVisibility(View.VISIBLE);
         btnChoose.setEnabled(false);
-        btnUpload.setEnabled(false);
+        btnUpload.setVisibility(View.GONE);
     }
 
     private void hideProgress() {
         mProgressBar.setVisibility(View.GONE);
         btnChoose.setEnabled(true);
-        btnUpload.setEnabled(true);
+        btnUpload.setVisibility(View.VISIBLE);
     }
 
     @NonNull
     private RequestBody createPartFromString(String descriptionString) {
-        return RequestBody.create(
-                okhttp3.MultipartBody.FORM, descriptionString);
+        return RequestBody.create(MediaType.parse(FileUtils.MIME_TYPE_TEXT), descriptionString);
     }
 
     @NonNull
     private MultipartBody.Part prepareFilePart(String partName, Uri fileUri) {
-        // https://github.com/iPaulPro/aFileChooser/blob/master/aFileChooser/src/com/ipaulpro/afilechooser/utils/FileUtils.java
         // use the FileUtils to get the actual file by uri
         File file = FileUtils.getFile(this, fileUri);
 
         // create RequestBody instance from file
-        RequestBody requestFile =
-                RequestBody.create(
-                        MediaType.parse(Objects.requireNonNull(getContentResolver().getType(fileUri))),
-                        file
-                );
+        RequestBody requestFile = RequestBody.create (MediaType.parse(FileUtils.MIME_TYPE_IMAGE), file);
 
         // MultipartBody.Part is used to send also the actual file name
         return MultipartBody.Part.createFormData(partName, file.getName(), requestFile);
@@ -234,80 +213,68 @@ public class MainActivity extends AppCompatActivity {
     /**
      *  Runtime Permission
      */
-    private boolean askForPermission() {
-        int currentAPIVersion = Build.VERSION.SDK_INT;
-        if (currentAPIVersion >= Build.VERSION_CODES.M) {
-            int hasCallPermission = ContextCompat.checkSelfPermission(MainActivity.this,
-                    Manifest.permission.READ_EXTERNAL_STORAGE);
-            if (hasCallPermission != PackageManager.PERMISSION_GRANTED) {
-                // Ask for permission
-                // need to request permission
-                if (ActivityCompat.shouldShowRequestPermissionRationale(MainActivity.this,
-                        Manifest.permission.READ_EXTERNAL_STORAGE)) {
-                    // explain
-                    showMessageOKCancel(
-                            new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface dialogInterface, int i) {
-                                    ActivityCompat.requestPermissions(MainActivity.this,
-                                            new String[]{android.Manifest.permission.READ_EXTERNAL_STORAGE},
-                                            REQUEST_CODE_ASK_PERMISSIONS);
-                                }
-                            });
-                    // if denied then working here
-                } else {
-                    // Request for permission
-                    ActivityCompat.requestPermissions(MainActivity.this,
-                            new String[]{android.Manifest.permission.READ_EXTERNAL_STORAGE},
-                            REQUEST_CODE_ASK_PERMISSIONS);
-                }
+    private void askForPermission() {
+        if ((ContextCompat.checkSelfPermission(this,
+                Manifest.permission.READ_EXTERNAL_STORAGE) +
+                ContextCompat.checkSelfPermission(this,
+                        Manifest.permission.WRITE_EXTERNAL_STORAGE))
+                != PackageManager.PERMISSION_GRANTED) {
+            /* Ask for permission */
+            // need to request permission
+            if (ActivityCompat.shouldShowRequestPermissionRationale(this,
+                    Manifest.permission.READ_EXTERNAL_STORAGE) ||
+                    ActivityCompat.shouldShowRequestPermissionRationale(this,
+                            Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
 
-                return false;
+                Snackbar.make(this.findViewById(android.R.id.content),
+                        "Please grant permissions to write data in sdcard",
+                        Snackbar.LENGTH_INDEFINITE).setAction("ENABLE",
+                        v -> ActivityCompat.requestPermissions(MainActivity.this,
+                                new String[]{Manifest.permission.READ_EXTERNAL_STORAGE,
+                                        Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                                REQUEST_CODE_PERMISSIONS)).show();
             } else {
-                // permission granted and calling function working
-                return true;
+                /* Request for permission */
+                ActivityCompat.requestPermissions(MainActivity.this,
+                        new String[]{Manifest.permission.READ_EXTERNAL_STORAGE,
+                                Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                        REQUEST_CODE_PERMISSIONS);
             }
+
         } else {
-            return true;
+            showChooser();
         }
     }
 
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        switch (requestCode) {
-            case REQUEST_CODE_ASK_PERMISSIONS:
-                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    // Permission Granted
-                    showChooser();
-                } else {
-                    // Permission Denied
-                    Toast.makeText(MainActivity.this, "Permission Denied!", Toast.LENGTH_SHORT).show();
-                }
-                break;
-            default:
-                super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == REQUEST_CODE_PERMISSIONS) {
+            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                // Permission Granted
+                showChooser();
+            } else {
+                // Permission Denied
+                Toast.makeText(MainActivity.this, "Permission Denied!", Toast.LENGTH_SHORT).show();
+            }
+        } else {
+            super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         }
     }
 
     private void showMessageOKCancel(DialogInterface.OnClickListener okListener) {
-
         AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
         final AlertDialog dialog = builder.setMessage("You need to grant access to Read External Storage")
                 .setPositiveButton("OK", okListener)
                 .setNegativeButton("Cancel", null)
                 .create();
 
-        dialog.setOnShowListener(new DialogInterface.OnShowListener() {
-            @Override
-            public void onShow(DialogInterface arg0) {
-                dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(
-                        ContextCompat.getColor(MainActivity.this, android.R.color.holo_blue_light));
-                dialog.getButton(AlertDialog.BUTTON_NEGATIVE).setTextColor(
-                        ContextCompat.getColor(MainActivity.this, android.R.color.holo_red_light));
-            }
+        dialog.setOnShowListener(arg0 -> {
+            dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(
+                    ContextCompat.getColor(MainActivity.this, android.R.color.holo_blue_light));
+            dialog.getButton(AlertDialog.BUTTON_NEGATIVE).setTextColor(
+                    ContextCompat.getColor(MainActivity.this, android.R.color.holo_red_light));
         });
 
         dialog.show();
-
     }
 }
